feat : Atualizando arquivos e estrutura de pastas
- Adicionado alguns arquivos dummies que serão implementados na próxima release.
- Adicionado os arquivos csv contendo os dados de propostas e comentários, para ser adicionado no banco na próxima release.
- Renomeado arquivo de análise de sentimentos.
- Alterado index e view principal para exibir um protótipo de análise de sentimento digitado pelo usuário.
