+++ requirements.txt
@@ -1,4 +1,3 @@
-requests>=2.32.0
 certifi>=2024.8.30
 openai>=1.56.1
 typing_extensions>=4.12.2
@@ -8,3 +7,4 @@ pydantic>=2.10.3
 streamlit>=1.40.0
 django>=5.1.4
 psycopg2-binary>=2.9.10
+requests>=2.32.0
